fix: duplicate imports and performance regression
- Remove duplication of imports in `util.py`
- Re-introduce `jit` decoration of `Kernel.compute`

While we wouldn't expect any runtime performance improvement from
explicitly calling `jit` on the `Kernel.compute` method, empirically
there is a strong degradation in test runtime when removing the JIT
decoration. It is presumed that the degradation is related to compile
time changes. Further investigation may be required in the future, but
for now it is sufficient to revert back to the explicit decoration.

refs: #566