feat: add coreset and coresubset data structures
As part of #581, it is necessary to decouple the coreset data-structures
and solvers/reduction-strategies. The new `coreax.coreset.Coreset` and
`coreax.coreset.Coresubset` data-structures introduced here facilitate
this decoupling, such that the conceptual notion of a coreset better
reflects our implementation. I.E. a coreset is a data-structure and a
solver is a solver, the current implementation treats a coreset as both
of these things.

Note: these new data-structures are currently unused. They will be used
in an upcoming PR to fully address #581.

A new set of unit tests have been created for these new data-structures
in `tests/unit/test_coreset.py`.

Additionally updates the new `coreax.data.Data` with `at_least{1,2}d`
equinox converters and a `__len__` magic method. This facilitates the
coreset `__len__` magic method, and will allow many `atleast_{1,2}d`
calls to be removed from the codebase, at a later date.

Tests for `coreax.data.Data` have been refactored to reduce code
duplication resulting from a new test for the new `__len__` method.

Refs: #518, #581