feat!: add `normalize` to `coreax.data.Data`
Adds the `coreax.data.Data.normalize` method - returns a copy of an
instance of `coreax.data.Data` with weights that sum to one.

Adds the `coreax.data.is_data` filter method - useful as an `is_leaf`
filter for `jax.tree_util` methods.

Also fixes an edge case where the default weights would be invalid if
the `data` parameter passed to `coreax.data.Data.__init__` was not 2D,
allowing the `eqx.field(converter=...)` attribute assignments to be
removed; the `jnp.atleast_2d` converter for `data` is now applied in
the `__init__` call; the `jnp.atleast_1d` converter for `weights` is
made redundant by the `jnp.broadcast_to(...)` call in `__init__`, which
itself made the `__check_init__` redundant as the broadcast will always
throw an error before the same invariant is checked in `__check_init_`.

Note: `__check_init__` in `coreax.data.SupervisedData` is still needed.

BREAKING CHANGE: `Data.weights` default to an un-normalized vector of
ones with length equal to `Data.data`; prior to this commit, the default
weights were normalized.