feat: add `compute_mean` to `coreax.kernel.Kernel`
The new `compute_mean` method generalizes the functionality of
`gramian_row_mean`, allowing any kernel matrix mean to be computed
block-wise. When data is weighted, the mean is computed in the same way
as `jax.numpy.average`.

In the future, the `compute_mean` method should be used to remove what
is now duplicated functionality from `coreax.metrics.MMD`.

Refs: #617