feat: replace performance test with jit_variants
Simplifies implementation of #581. The performance tests used to check
if JIT compilation significantly reduced the runtime of kernel methods
and some coreset algorithms. Significance was determined by a KS test
between the distribution of repeated pre and post JIT run-times.

While knowing if JIT compilation enhances performance is useful, it is
not a strict property of our methods. The critical check is that the
methods can be successfully JIT compiled (a now necessary property).

The impact of JIT compilation on runtime performance should be covered
by a future pull to address #509 (performance benchmarking).

Refs: #509, #581