fix: correct refinement methods
Refinement methods have been updated to correctly handle zero weighted
indices and data. This allows us to pad with zero weighted data, to
preserve shape stability, without altering the result of the algorithm
in any way. A new problem case, "random-zero-weights", and a new
invariant check have been added to the solver tests to check this
behaviour.

The refinement methods have also been updated to correctly handle
initial coresets. The kernel similarity penalty is now computed from the
initialisation, rather than defaulting to the zero vector, ensuring that
the result is not initialisation invariant (as it was before).

Additionally improves the clarity of presentation for the `SteinThinning`
selection function.