feat: enable jaxtyped `Data` and `SupervisedData`.
Enables runtime type checking of the inputs to `Data` and
`SupervisedData`. To facilitate this, modifications have been made
to facilitate this, some type-hints have been updated, and the
functionality of the internal `_atleast_2d_consistent` function
has been simplified (now only operating on single arrays not
sequences of arrays; the later functionality was unused).