feat: remove `SilentTQDM`. (#1016)
* feat: remove `SilentTQDM`.

To reduce the number of lines of code, and subsequent complexity
throughout the codebase, the `SilentTQDM` class has been removed
without loss of functionality.

The `SilentTQDM` functionality is now provided through a boolean
`progress_bar` variable and silencing performed via the `disable`
keyword argument to `tqdm.tqdm`. The `write` method is silenced
by making all calls to the method conditional on the value of the
`progress_bar` boolean.

* chore(docs): update changelog.

Add the deprecation of `SilentTQDM` of the changelog.