fix: decorate abstract fixtures.
Many tests classes have abstract methods that are meant to be
implemented as fixtures but are not decorated with `pytest.fixture`.
By adding these missing decorators, we placate pyright and provide a
clearer indication to future developers of the intent for concrete
implementations, beyond what is decsribe in the docstrings.