fix: upgrade type hints.
By dropping support for Python 3.9, it is possible to replace the
`Union` and `Optional` type hints with the `|` operator. At the
same time, it is also possible to simplify some of the jaxtyped
type hints.

BREAKING CHANGES: Sequence support for `_atleast_2d_consistent`
has been dropped. The functionality was unused and significantly
complicated the type hints.