chore(ci): test lowest resolution. (#1092)
Add the `uv` resolution mode to the unittest.yml matrix. Currently tests
the `highest` and `lowest` resolution modes to provide coverage
of the lower bounds specified in the pyproject.

A set of constraint-dependencies have been specified to facilitate this.
The lower bounds of certain dev and test tools have been bumped as a
matter of good practice.

Refs: #918